refactor: move management of targets (intermediate)
